working 99% will continue to work on it
I'm going to go ahead and post this to see if it work with your tables but for me, I don't know why but it run fine until adding parent to student 48 and 50 referring might be a syntax error on my part but read through it and could not find it. anyway I'll keep working on it on my end.